Add fix for crash with unnoticed QSPI commands
If a QSPI command is missed, the number of clock cycles in the
transaction may not be a multiple of eight. That can cause a ValueError
when it starts decoding the next command because the `_mosi_out` and
`_miso_in` were not zeroed between transactions. This adds the zeroing
to fix this crash, at the expense of missing the last byte(s) in the
prior transaction.